Add record functionality

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package edu.miu.cs.cs544.repository;
 import edu.miu.common.repository.BaseRepository;
 import edu.miu.cs.cs544.domain.Member;
 
-public interface MemberRepository extends BaseRepository<Member, Integer>{
+public interface MemberRepository extends BaseRepository<Member, Long>{
 
     Member findById(Long memberId);
 }
